+++ dto/transacao_request.go
@@ -1,7 +1,7 @@
 package dto
 
 type TransacaoRequest struct {
-	Valor     float64 `json:"valor"`
-	Tipo      string  `json:"tipo"`
-	Descricao string  `json:"descricao"`
+	Valor     int64  `json:"valor"`
+	Tipo      string `json:"tipo"`
+	Descricao string `json:"descricao"`
 }
