+++ entities/cliente.go
@@ -4,7 +4,7 @@ import "github.com/uptrace/bun"
 
 type Cliente struct {
 	bun.BaseModel `bun:"table:clientes,alias:u"`
-	ID            string  `json:"id" bun:"id_cliente,pk,autoincrement"`
-	Saldo         float64 `json:"saldo"`
-	Limite        float64 `json:"limite"`
+	ID            string `json:"id" bun:"id_cliente,pk,autoincrement"`
+	Saldo         int64  `json:"saldo"`
+	Limite        int64  `json:"limite"`
 }
